Create Workbok
BMI is a small project developed by using Python language. And is a measure of relative 
weight based on an individual’s mass and height. Today, Body Mass Index is commonly        
used to classify people as underweight, overweight, and even with obesity. Also, it is 
adopted by countries to promote healthy eating